Partiel Admin Console
This script renders just a few of the pages for the admin console. I'm
wanted to work more on the Server Software before I finished this script
but at least you can tell me if the details are clear enough. It
requires redemptionweb (which I made) which I'll give you as soon as I
publish it under MIT which will be today sometime.